Move the mock datetime utility to a mocks package.
