+++ tests/testsupport/mock/builtins.py
